requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
diff-match-patch==20230430
Django==5.1
django-admin-material-dashboard==1.0.10
django-allauth==64.1.0
django-cors-headers==4.4.0
django-import-export==4.1.1
django-unfold==0.36.0
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pyOpenSSL==24.2.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
setuptools==72.2.0
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
waitress==3.0.0
whitenoise==6.7.0
